day3

+++ src/day1.h
@@ -1,7 +1,7 @@
 #ifndef AOC25_SRC_DAY1_H
 #define AOC25_SRC_DAY1_H
 
-int day1Part1(char inputFilePath[]);
-int day1Part2(char inputFilePath[]);
+long day1Part1(char inputFilePath[]);
+long day1Part2(char inputFilePath[]);
 
 #endif
